+++ ChangeLog.md
@@ -12,6 +12,7 @@ Behavior changes:
 Other enhancements:
 
 * Travis no longer builds using `sudo: false` as that behaviour is due to be deprecated.
+* Add `--cabal-files` flag to `stack ide targets` command.
 
 Bug fixes: